% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{dijkstra}
\alias{dijkstra}
\title{Runs Dijkstra algorithm}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{A dataframe.}

\item{init_node}{A numeric scalar that exists in the graph.}
}
\value{
A vector
}
\description{
Runs Dijkstra algorithm
}
\examples{
wiki_graph <- data.frame(
v1=c(1,1,1,2,2,2,3,3,3,3,4,4,4,5,5,6,6,6),
v2=c(2,3,6,1,3,4,1,2,4,6,2,3,5,4,6,1,3,5),
w=c(7,9,14,7,10,15,9,10,11,2,15,11,6,6,9,14,2,9)
)
dijkstra(wiki_graph, 1)
dijkstra(wiki_graph, 3)
}
\references{
https://en.wikipedia.org/wiki/Dijkstra%27s_algorithm
}
