% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{euclidean}
\alias{euclidean}
\title{Finds greatest common divisor using euclidean algorithm.}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{An integer.}

\item{b}{An integer.}
}
\value{
An integer.
}
\description{
Finds greatest common divisor using euclidean algorithm.
}
\examples{
euclidean(123612, 13892347912)
euclidean(100, 1000)
}
\references{
https://en.wikipedia.org/wiki/Euclidean_algorithm
}
