% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heleramcar3-package.R
\docType{package}
\name{heleramcar3-package}
\alias{heleramcar3}
\alias{heleramcar3-package}
\title{heleramcar3: Euclidean and Dijkstra algorithms for Computar Lab 3}
\description{
Implementation of Euclidean and Dijkstra algorithms for Advanced R course in Linköping University.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ramonamezquita/heleramcar3}
}

}
\author{
\strong{Maintainer}: Helena Llorens \email{hllor282@student.liu.se}

Authors:
\itemize{
  \item Carmen Rodríguez \email{carro162@student.liu.se}
  \item Ramón Amézquita \email{ramam120@student.liu.se}
}

}
\keyword{internal}
