% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Wikipedia graph data}
\format{
A data frame with 18 rows and 3 columns:
\describe{
  \item{v1}{First edge of the graph from v1 to v2}
  \item{v2}{Second edge of the graph from v1 to v2}
  \item{w}{The weight of the edge}
  ...
}
}
\source{
<https://en.wikipedia.org/wiki/Dijkstra's_algorithm>
}
\usage{
wiki_graph
}
\description{
First graph at the Wikipedia page
}
\keyword{datasets}
